Move dynamo client library indirection to separate file.
